+++ Tests/PbxprojTests/Helper.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Pbxproj
 
 func pathForFixture(fileName: String) -> String {
     #if Xcode
@@ -7,4 +8,13 @@ func pathForFixture(fileName: String) -> String {
     #else
         return "Tests/AsciiPlistParserTests/Fixtures/\(fileName)"
     #endif
+}
+
+func _pbxproj() -> Pbxproj {
+    #if Xcode
+        let path = pathForFixture(fileName: "test.pbxproj")
+    #else
+        let path = pathForFixture(fileName: "Xcode/8.3.2/SingleViewApplication/SingleViewApplication.xcodeproj/project.pbxproj")
+    #endif
+    return try! Pbxproj(path: path)
 }
